Color in chart is per column, not global.
I made a mistake using renewables for the chart color,
but this value is global and will change the whole chart.
What I want is that if a column is lower than 50%, it should be
red. But, because the % to pixel conversion, numbers such as 49.80%
might render a LED red and a chart column green.